refactor: Fix circular dependency error in emailverification recipe and fix tests failures
