+++ supertokens_python/recipe/passwordless/recipe.py
@@ -145,8 +145,9 @@ class PasswordlessRecipe(RecipeModule):
         )
 
         def callback():
-            ev_recipe = EmailVerificationRecipe.get_instance()
-            ev_recipe.add_get_email_for_user_id_func(self.get_email_for_user_id)
+            ev_recipe = EmailVerificationRecipe.get_instance_optional()
+            if ev_recipe:
+                ev_recipe.add_get_email_for_user_id_func(self.get_email_for_user_id)
 
         PostSTInitCallbacks.add_post_init_callback(callback)
