+++ tests/emailpassword/test_emaildelivery.py
@@ -36,6 +36,7 @@ from supertokens_python.recipe.emailpassword import (
     InputResetPasswordUsingTokenFeature,
 )
 from supertokens_python.recipe.emailpassword.emaildelivery.services import SMTPService
+from supertokens_python.recipe.emailverification.emaildelivery.services import SMTPService as EVSMTPService
 from supertokens_python.recipe.emailpassword.types import (
     EmailTemplateVars,
     PasswordResetEmailTemplateVars,
@@ -526,7 +527,11 @@ async def test_email_verification_default_backward_compatibility(
             api_base_path="/auth",
         ),
         framework="fastapi",
-        recipe_list=[emailpassword.init(), session.init()],
+        recipe_list=[
+            emailverification.init(ParentRecipeEmailVerificationConfig(mode="OPTIONAL")),
+            emailpassword.init(),
+            session.init()
+        ],
     )
     start_st()
 
@@ -587,7 +592,11 @@ async def test_email_verification_default_backward_compatibility_suppress_error(
             api_base_path="/auth",
         ),
         framework="fastapi",
-        recipe_list=[emailpassword.init(), session.init()],
+        recipe_list=[
+            emailverification.init(ParentRecipeEmailVerificationConfig(mode="OPTIONAL")),
+            emailpassword.init(),
+            session.init()
+        ],
     )
     start_st()
 
@@ -698,11 +707,11 @@ async def test_email_verification_custom_override(driver_config_client: TestClie
     email = ""
     email_verify_url = ""
 
-    def email_delivery_override(oi: EmailDeliveryInterface[EmailTemplateVars]):
+    def email_delivery_override(oi: EmailDeliveryInterface[VerificationEmailTemplateVars]):
         oi_send_email = oi.send_email
 
         async def send_email(
-            template_vars: EmailTemplateVars, user_context: Dict[str, Any]
+            template_vars: VerificationEmailTemplateVars, user_context: Dict[str, Any]
         ):
             nonlocal email, email_verify_url
             email = template_vars.user.email
@@ -723,12 +732,16 @@ async def test_email_verification_custom_override(driver_config_client: TestClie
         ),
         framework="fastapi",
         recipe_list=[
-            emailpassword.init(
-                email_delivery=EmailDeliveryConfig(
-                    service=None,
-                    override=email_delivery_override,
+            emailverification.init(
+                ParentRecipeEmailVerificationConfig(
+                    mode="OPTIONAL",
+                    email_delivery=EmailDeliveryConfig(
+                        service=None,
+                        override=email_delivery_override,
+                    )
                 )
             ),
+            emailpassword.init(),
             session.init(),
         ],
     )
@@ -782,7 +795,7 @@ async def test_email_verification_smtp_service(driver_config_client: TestClient)
         False,
     )
 
-    def smtp_service_override(oi: SMTPServiceInterface[EmailTemplateVars]):
+    def smtp_service_override(oi: SMTPServiceInterface[VerificationEmailTemplateVars]):
         async def send_raw_email_override(
             content: EmailContent, _user_context: Dict[str, Any]
         ):
@@ -796,7 +809,7 @@ async def test_email_verification_smtp_service(driver_config_client: TestClient)
             # Note that we aren't calling oi.send_raw_email. So Transporter won't be used.
 
         async def get_content_override(
-            template_vars: EmailTemplateVars, _user_context: Dict[str, Any]
+            template_vars: VerificationEmailTemplateVars, _user_context: Dict[str, Any]
         ) -> EmailContent:
             nonlocal get_content_called, email_verify_url
             get_content_called = True
@@ -816,7 +829,7 @@ async def test_email_verification_smtp_service(driver_config_client: TestClient)
 
         return oi
 
-    email_delivery_service = SMTPService(
+    email_delivery_service = EVSMTPService(
         smtp_settings=SMTPSettings(
             host="",
             from_=SMTPSettingsFrom("", ""),
@@ -828,12 +841,12 @@ async def test_email_verification_smtp_service(driver_config_client: TestClient)
     )
 
     def email_delivery_override(
-        oi: EmailDeliveryInterface[EmailTemplateVars],
-    ) -> EmailDeliveryInterface[EmailTemplateVars]:
+        oi: EmailDeliveryInterface[VerificationEmailTemplateVars],
+    ) -> EmailDeliveryInterface[VerificationEmailTemplateVars]:
         oi_send_email = oi.send_email
 
         async def send_email_override(
-            template_vars: EmailTemplateVars, user_context: Dict[str, Any]
+            template_vars: VerificationEmailTemplateVars, user_context: Dict[str, Any]
         ):
             nonlocal outer_override_called
             outer_override_called = True
@@ -852,12 +865,16 @@ async def test_email_verification_smtp_service(driver_config_client: TestClient)
         ),
         framework="fastapi",
         recipe_list=[
-            emailpassword.init(
-                email_delivery=EmailDeliveryConfig(
-                    service=email_delivery_service,
-                    override=email_delivery_override,
+            emailverification.init(
+                ParentRecipeEmailVerificationConfig(
+                    mode="OPTIONAL",
+                    email_delivery=EmailDeliveryConfig(
+                        service=email_delivery_service,
+                        override=email_delivery_override,
+                    )
                 )
             ),
+            emailpassword.init(),
             session.init(),
         ],
     )
